fix(post): 특정 이메일로 게시글 조회 및 댓글 좋아요 회원정보 연동

+++ src/main/java/com/github/commerce03/service/LikeService.java
@@ -4,6 +4,9 @@ import com.github.commerce03.repository.commend.Commend;
 import com.github.commerce03.repository.commend.CommendRepository;
 import com.github.commerce03.repository.like.Like;
 import com.github.commerce03.repository.like.LikeJpaRepository;
+import com.github.commerce03.repository.user.UserEntity;
+import com.github.commerce03.repository.user.UserJpaRepository;
+import com.github.commerce03.service.exeptions.NotFoundException;
 import com.github.commerce03.service.mapper.LikeMapper;
 import com.github.commerce03.web.dto.like.LikeRequestDto;
 import lombok.RequiredArgsConstructor;
@@ -19,13 +22,20 @@ public class LikeService {
     private final LikeJpaRepository likeJpaRepository;
     private final CommendRepository commendRepository;
 
+    private final UserJpaRepository userJpaRepository;
+
     @Transactional
     public String commendLike(LikeRequestDto likeRequestDto) {
         Integer commendId = likeRequestDto.getCommendId();
-        Optional<Like> like = likeJpaRepository.findByCommend_ComId(commendId);
+        String userEmail = likeRequestDto.getUserEmail();
+
+        UserEntity user = userJpaRepository.findByUserEmail(userEmail);
+        if(user==null)throw new NotFoundException("유저정보를 찾을수 없습니다.");
+
+        Optional<Like> like = likeJpaRepository.findByCommend_ComIdAndUser(commendId,user);
         if(like.isEmpty()){
             Commend commend = commendRepository.findById(commendId).orElseThrow(()->new RuntimeException("댓글이 존재하지 않습니다."));
-            likeJpaRepository.save(LikeMapper.INSTANCE.LikeRequestDtoToLike(commend));
+            likeJpaRepository.save(LikeMapper.INSTANCE.LikeRequestDtoToLike(commend,user));
         }else{
             likeJpaRepository.delete(like.get());
         }
